+++ lib/mongoid/includes/association/eager.rb
@@ -1,6 +1,6 @@
 module Mongoid
   module Includes
-    module Relations
+    module Association
 
       # Internal: Patch to the base class for eager load preload functions,
       # that interacts with the Mongoid::Includes::Inclusion as @medatada.
@@ -9,7 +9,7 @@ module Mongoid
         # Internal: Performs eager load and iterates over the preloaded documents
         # for the current relation.
         def each_loaded_document
-          @metadata.load_documents_for(key, keys_from_docs).each do |doc|
+          @association.load_documents_for(key, keys_from_docs).each do |doc|
             yield doc
           end
         end