+++ spec/support/models/address.rb
@@ -26,44 +26,8 @@ class Address
 
   belongs_to :band
 
-  embedded_in :addressable, polymorphic: true do
-    def extension
-      "Testing"
-    end
-    def doctor?
-      title == "Dr"
-    end
-  end
-
-  scope :without_postcode, ->{ where(postcode: nil) }
-  scope :rodeo, ->{ where(street: "Rodeo Dr") } do
-    def mansion?
-      all? { |address| address.street == "Rodeo Dr" }
-    end
-  end
+  embedded_in :addressable, polymorphic: true
 
   validates_presence_of :street, on: :update
   validates_format_of :street, with: /\D/, allow_nil: true
-
-  def set_parent=(set = false)
-    self.parent_title = addressable.title if set
-  end
-
-  def <=>(other)
-    street <=> other.street
-  end
-
-  class << self
-    def california
-      where(state: "CA")
-    end
-
-    def homes
-      where(address_type: "Home")
-    end
-
-    def streets
-      all.map(&:street)
-    end
-  end
 end