+++ spec/support/models/post.rb
@@ -1,3 +1,5 @@
+require_relative './person'
+
 class Post
   include Mongoid::Document
   include Mongoid::Attributes::Dynamic